Add files via upload
This project utilizes the "tabularisai/multilingual-sentiment-analysis" model from Hugging Face to perform sentiment classification on text in multiple languages. The model predicts whether a given text is positive, negative, or neutral.
here i am accessing the HuggingFace and chatgpt product key and it's direct to LLM in hugging face tool.